requirements.txt
pymilvus>=2.6.0
anthropic>=0.90.0
google-genai>=1.0.0
chonkie[semantic]>=1.0.0
liteparse>=1.0.0
